Add IMigrationTesterService and implementations for SQL and EF migrations
