+++ test/Infrastructure.IntegrationTest/DatabaseDataAttribute.cs
@@ -1,22 +1,15 @@
-﻿using System.Data;
-using System.Reflection;
+﻿using System.Reflection;
 using Bit.Core.Enums;
 using Bit.Core.Settings;
 using Bit.Infrastructure.Dapper;
 using Bit.Infrastructure.EntityFramework;
 using Bit.Infrastructure.EntityFramework.Repositories;
-using Bit.Migrator;
-using Microsoft.Data.Sqlite;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Infrastructure;
-using Microsoft.EntityFrameworkCore.Migrations;
+using Bit.Infrastructure.IntegrationTest.Services;
 using Microsoft.Extensions.Caching.Distributed;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Time.Testing;
-using MySqlConnector;
-using Npgsql;
 using Xunit.Sdk;
 
 namespace Bit.Infrastructure.IntegrationTest;
@@ -124,83 +117,15 @@ public class DatabaseDataAttribute : DataAttribute
 
     private void AddSqlMigrationTester(IServiceCollection services, string connectionString, string migrationName)
     {
-        services.AddSingleton<IMigrationTester, SqlMigrationTester>(sp => new SqlMigrationTester(connectionString, migrationName));
+        services.AddSingleton<IMigrationTesterService, SqlMigrationTesterService>(sp => new SqlMigrationTesterService(connectionString, migrationName));
     }
 
     private void AddEfMigrationTester(IServiceCollection services, SupportedDatabaseProviders databaseType, string migrationName)
     {
-        services.AddSingleton<IMigrationTester, EfMigrationTester>(sp =>
+        services.AddSingleton<IMigrationTesterService, EfMigrationTesterService>(sp =>
         {
             var dbContext = sp.GetRequiredService<DatabaseContext>();
-            return new EfMigrationTester(dbContext, databaseType, migrationName);
+            return new EfMigrationTesterService(dbContext, databaseType, migrationName);
         });
-    }
-}
-
-public interface IMigrationTester
-{
-    void ApplyMigration();
-}
-
-public class SqlMigrationTester : IMigrationTester
-{
-    private readonly string _connectionString;
-    private readonly string _migrationName;
-
-    public SqlMigrationTester(string connectionString, string migrationName)
-    {
-        _connectionString = connectionString;
-        _migrationName = migrationName;
-    }
-
-    public void ApplyMigration()
-    {
-        var dbMigrator = new DbMigrator(_connectionString);
-        dbMigrator.MigrateMsSqlDatabaseWithRetries(scriptName: _migrationName, repeatable: true);
-    }
-}
-
-public class EfMigrationTester : IMigrationTester
-{
-    private readonly DatabaseContext _databaseContext;
-    private readonly SupportedDatabaseProviders _databaseType;
-    private readonly string _migrationName;
-
-    public EfMigrationTester(DatabaseContext databaseContext, SupportedDatabaseProviders databaseType, string migrationName)
-    {
-        _databaseContext = databaseContext;
-        _databaseType = databaseType;
-        _migrationName = migrationName;
-    }
-
-    public void ApplyMigration()
-    {
-        // Delete the migration history to ensure the migration is applied
-        DeleteMigrationHistory();
-
-        var migrator = _databaseContext.GetService<IMigrator>();
-        migrator.Migrate(_migrationName);
-    }
-
-    private void DeleteMigrationHistory()
-    {
-        var deleteCommand = "DELETE FROM __EFMigrationsHistory WHERE MigrationId LIKE @migrationName";
-        IDbDataParameter? parameter = null;
-
-        switch (_databaseType)
-        {
-            case SupportedDatabaseProviders.MySql:
-                parameter = new MySqlParameter("@migrationName", "%" + _migrationName);
-                break;
-            case SupportedDatabaseProviders.Postgres:
-                deleteCommand = "DELETE FROM \"__EFMigrationsHistory\" WHERE \"MigrationId\" LIKE @migrationName";
-                parameter = new NpgsqlParameter("@migrationName", "%" + _migrationName);
-                break;
-            case SupportedDatabaseProviders.Sqlite:
-                parameter = new SqliteParameter("@migrationName", "%" + _migrationName);
-                break;
-        }
-
-        _databaseContext.Database.ExecuteSqlRaw(deleteCommand, parameter);
     }
 }
