+++ test/Infrastructure.IntegrationTest/Services/IMigrationTesterService.cs
@@ -1,0 +1,6 @@
+﻿namespace Bit.Infrastructure.IntegrationTest.Services;
+
+public interface IMigrationTesterService
+{
+    void ApplyMigration();
+}
